requirements.txt
fastapi>=0.115,<1.0
docxtpl
uvicorn[standard]>=0.30,<1.0
python-multipart>=0.0.9,<1.0
python-docx>=1.1,<2.0
pypdf>=4.0,<6.0
google-genai>=1.0
jinja2>=3.1,<4.0
pillow>=10.0,<12.0
httpx>=0.27,<1.0